Add CentOS support to ruby cookbooks
Since we're pre-compiling Ruby, we need to compile a version for each
distro family, as libraries are different versions and in different
places so things might break otherwise. The change to the Jenkins job to
support this was minimal, thanks to Docker.

This also changes the kitchen configuration slightly to stop running
tests as `root`. This became a problem on CentOS, where root's PATH
seems to be immutable (there were a couple threads on-line where I saw
people having similar issues). Running tests as a normal user will be
fine, since that's where most of the stuff we care about will live
anyway.